Split generation into one-per-type.
That is, we now have a generated .cpp and .hpp per-type, along
with a generated ros2_topics.hpp that knows how to access all
types.

The way it works is that the per-type .hpp file contains a factory
function for pubs and subs, which are "registered" with ros2_topics.hpp
at generation time.  During runtime, each topic of a type is looked
up in the type->factory map, and an appropriate object of that type
is created.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>