Fix looping forever bug.
transporter->read() is a blocking call, for the amount of
time specified in the poll_ms variable when the class is
constructed.  If a writer were to keep writing at a rate
shorter than the poll time, then the inner while loop would
never quit, since there would always be data available.
Instead, change this to an if statement so we only process
one message before checking the running variable.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>